text-synthesis/requirements.txt
## The following requirements were added by pip freeze:                                                                                                                                           
aiohappyeyeballs==2.4.3                                                                                                                                                                           
aiohttp==3.10.10                                                                                                                                                                                  
aiosignal==1.3.1                                                                                                                                                                                  
async-timeout==4.0.3                                                                                                                                                                              
attrs==24.2.0                                                                                                                                                                                     
audioread==3.0.1                                                                                                                                                                                  
cachetools==5.5.0                                                                                                                                                                                 
certifi==2024.8.30                                                                                                                                                                                
cffi==1.17.1                                                                                                                                                                                      
charset-normalizer==3.4.0                                                                                                                                                                         
click==8.1.7                                                                                                                                                                                      
contourpy==1.3.0                                                                                                                                                                                  
cycler==0.12.1                                                                                                                                                                                    
datasets==2.21.0                                                                                                                                                                                  
decorator==5.1.1                                                                                                                                                                                  
dill==0.3.8                                                                                                                                                                                       
exceptiongroup==1.2.2                                                                                                                                                                             
filelock==3.16.1                                                                                                                                                                                  
fonttools==4.54.1                                                                                                                                                                                 
frozenlist==1.4.1                                                                                                                                                                                 
fsspec==2024.6.1                                                                                                                                                                                  
google-api-core==2.21.0                                                                                                                                                                           
google-auth==2.35.0                                                                                                                                                                               
google-cloud-core==2.4.1                                                                                                                                                                          
google-cloud-translate==3.16.0                                                                                                                                                                    
googleapis-common-protos==1.65.0                                                                                                                                                                  
grpc-google-iam-v1==0.13.1                                                                                                                                                                        
grpcio==1.67.0                                                                                                                                                                                    
grpcio-status==1.67.0
huggingface-hub==0.26.0
HyperPyYAML==1.2.2
hypothesis==6.115.3
idna==3.10
Jinja2==3.1.4
jiwer==2.6.0
joblib==1.4.2
kiwisolver==1.4.7
librosa==0.9.2
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.1
noisereduce==3.0.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyctcdecode==0.5.0
pygtrie==2.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==2.13.7
regex==2024.9.11
requests==2.32.3
resampy==0.4.3
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
speechbrain==1.0.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
torchaudio==2.5.0
tqdm==4.66.5
transformers==4.45.2
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.15.5